fix: renamed conter to counter
