+++ src/services/extensions/counter.ts
